試験的マス目外歩行
試験的です

+++ kyoudoukikaku/player.pde
@@ -25,7 +25,7 @@ class player extends people{//プレイヤー
      attackanim[1]=attackanim[2];
      attackanim[2]=m;
   }
-  void draw(int _x,int _y){//座標指定あり描画
+  void draw(float _x,float _y){//座標指定あり描画
   int attackd=0;
     fcount++;
     if(attacking){
